templates/requirements.txt
Flask==2.0.1
google-generativeai==1.0.0
